requirements.txt
pillow
pymongo
python-bsonjs
flask
wtform
flask-markdown
flask-wtf
werkzeug